Storage refactor and test fixes
