+++ modules/json_form_widget/src/SchemaUiHandler.php
@@ -3,6 +3,7 @@
 namespace Drupal\json_form_widget;
 
 use Drupal\Core\DependencyInjection\ContainerInjectionInterface;
+use Drupal\metastore\SchemaRetriever;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Drupal\Core\Logger\LoggerChannelFactory;
 use Drupal\metastore\SchemaRetrieverInterface;
