+++ modules/metastore/modules/metastore_search/tests/src/Unit/Plugin/search_api/datasource/DkanDatasetTest.php
@@ -7,7 +7,7 @@ use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\Entity\EntityTypeManager;
 use Drupal\Core\Entity\Query\QueryInterface;
 use Drupal\metastore\Storage\AbstractEntityStorage;
-use Drupal\metastore\Storage\NodeStorageFactory;
+use Drupal\metastore\Storage\MetastoreDataNodeStorageFactory;
 use Drupal\metastore_search\Plugin\search_api\datasource\DkanDataset;
 use Drupal\node\NodeInterface;
 use MockChain\Chain;
@@ -15,6 +15,8 @@ use MockChain\Options;
 use MockChain\Sequence;
 use PHPUnit\Framework\TestCase;
 use Drupal\Core\Entity\EntityTypeRepository;
+use Drupal\metastore\MetastoreDataNode;
+use Drupal\metastore\Storage\MetastoreDataNodeStorage;
 use Drupal\metastore_search\ComplexData\Dataset;
 
 /**
@@ -32,7 +34,7 @@ class DkanDatasetTest extends TestCase {
     $containerOptions = (new Options())
       ->add('entity_type.manager', EntityTypeManager::class)
       ->add('entity_type.repository', EntityTypeRepository::class)
-      ->add('dkan.metastore.storage', NodeStorageFactory::class)
+      ->add('dkan.metastore.storage', MetastoreDataNodeStorageFactory::class)
       ->index(0);
 
     $nids = [1, 2];
@@ -48,9 +50,9 @@ class DkanDatasetTest extends TestCase {
       ->add(QueryInterface::class, 'execute', $executeSequence)
       ->add(QueryInterface::class, 'range', QueryInterface::class)
       ->add(EntityTypeRepository::class, 'getEntityTypeFromClass', NULL)
-      ->add(NodeStorageFactory::class, 'getInstance', AbstractEntityStorage::class)
-      ->add(AbstractEntityStorage::class, 'retrieve', '{}')
-      ->add(DataFactory::class, 'getInstance', Data::class)
+      ->add(MetastoreDataNodeStorageFactory::class, 'getInstance', MetastoreDataNodeStorage::class)
+      ->add(NodeStorage::class, 'retrieve', '{}')
+      ->add(NodeStorage::class, 'getEntityIdFromUuid', 1)
       ->add(Data::class, 'retrievePublished', '{}')
       ->getMock();
