--- modules/metastore/src/Storage/NodeStorage.php
@@ -1,62 +1,0 @@
-<?php
-
-namespace Drupal\metastore\Storage;
-
-use Drupal\Core\Entity\EntityTypeManager;
-
-/**
- * Node Data.
- */
-class NodeStorage extends AbstractEntityStorage {
-
-  /**
-   * NodeData constructor.
-   */
-  public function __construct(string $schemaId, EntityTypeManager $entityTypeManager) {
-    $this->entityType = 'node';
-    parent::__construct($schemaId, $entityTypeManager);
-    $this->bundle = 'data';
-  }
-
-  /**
-   * Inherited.
-   *
-   * {@inheritdoc}.
-   */
-  public function retrieve(string $uuid) : ?string {
-
-    if ($this->getDefaultModerationState() === 'published') {
-      $entity = $this->getEntityPublishedRevision($uuid);
-    }
-    else {
-      $entity = $this->getEntityLatestRevision($uuid);
-    }
-
-    if ($entity) {
-      return $entity->get('field_json_metadata')->getString();
-    }
-
-    throw new \Exception("No data with that identifier was found.");
-  }
-
-  /**
-   * Get the entity id from the dataset identifier.
-   *
-   * @param string $uuid
-   *   The dataset identifier.
-   *
-   * @return int|null
-   *   The entity id, if found.
-   */
-  public function getEntityIdFromUuid(string $uuid) : ?int {
-
-    $entity_ids = $this->entityStorage->getQuery()
-      ->condition('uuid', $uuid)
-      ->condition($this->bundleKey, $this->bundle)
-      ->condition('field_data_type', $this->schemaId)
-      ->execute();
-
-    return $entity_ids ? (int) reset($entity_ids) : NULL;
-  }
-
-}
